Moved suggestion bar to own component
